adds tests for navigation on all existing pages
