+++ src/routes/+layout.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import Nav from '$components/Nav.svelte';
+	import Nav from '$components/Nav/Nav.svelte';
   import './global.css';
 </script>
