+++ src/routes/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import ParallaxForest from '$components/scenes/ParallaxForest.svelte';
   import Rain from '$components/Rain.svelte';
-	import Nav from '$components/Nav.svelte';
+	import Nav from '$components/Nav/Nav.svelte';
 </script>
 
 <ParallaxForest />
