+++ src/routes/global.css
@@ -24,12 +24,12 @@ body {
   background: var(--dark-500);
   overflow: hidden;
   box-sizing: border-box;
+}
 
-  * {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-  }
+body * {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
 }
 
 h1 {
